Implemented a new way to poll for the Git config rotator.
The plugin no longer requres a workspace to poll. It will look at a previous valid configuration and detect if the head of the configured branch looks different. If it is, there are changes.

The actual 'find next commit' action is done in the checkout phase where the tree is traversed and the next commit determined.